fix multi-email issue
